+++ gdpr/src/main/java/com/kairos/controller/data_inventory/asset/OrganizationStorageFormatController.java
@@ -43,7 +43,7 @@ public class OrganizationStorageFormatController {
 
     @ApiOperation("get StorageFormat by id")
     @GetMapping("/storage_format/{storageFormatId}")
-    public ResponseEntity<Object> getStorageFormat(@PathVariable Long unitId, @PathVariable BigInteger storageFormatId) {
+    public ResponseEntity<Object> getStorageFormat(@PathVariable Long unitId, @PathVariable Long storageFormatId) {
         return ResponseHandler.generateResponse(HttpStatus.OK, true, storageFormatService.getStorageFormat(unitId, storageFormatId));
     }
 
@@ -65,7 +65,7 @@ public class OrganizationStorageFormatController {
 
     @ApiOperation("update StorageFormat by id")
     @PutMapping("/storage_format/{storageFormatId}")
-    public ResponseEntity<Object> updateStorageFormat(@PathVariable Long unitId, @PathVariable BigInteger storageFormatId, @Valid @RequestBody StorageFormatDTO storageFormat) {
+    public ResponseEntity<Object> updateStorageFormat(@PathVariable Long unitId, @PathVariable Long storageFormatId, @Valid @RequestBody StorageFormatDTO storageFormat) {
         return ResponseHandler.generateResponse(HttpStatus.OK, true, storageFormatService.updateStorageFormat(unitId, storageFormatId, storageFormat));
 
     }
